minor improvements
* Add callback to RunThrotlle
* Add config for Sdk logging
* Move Cognite utils classes to own files
